app/src/main/java/com/example/mystoryapp2/data/viewmodel/ListStoryViewModel.kt
package com.example.mystoryapp2.data.viewmodel

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.example.mystoryapp2.data.model.ListStoryItem
import com.example.mystoryapp2.data.paging.StoryRepository

class ListStoryViewModel(storyRepository: StoryRepository): ViewModel() {
    val stories: LiveData<PagingData<ListStoryItem>> = storyRepository.getStory().cachedIn(viewModelScope)
}